feat: 학생코드 연결

+++ dodam-system-domain/dodam-domain-rds/src/main/java/b1nd/dodam/domain/rds/member/repository/ParentRepository.java
@@ -3,6 +3,6 @@ package b1nd.dodam.domain.rds.member.repository;
 import b1nd.dodam.domain.rds.member.entity.Parent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ParentRepository extends JpaRepository<Parent, Long> {
+public interface ParentRepository extends JpaRepository<Parent, Integer> {
 
 }
